requirements.txt
psycopg2-binary
sqlalchemy-cockroachdb
sqlalchemy
FLASK==1.1.1 